Rename class files and move to roof subfolder
